# Simplify code by precomputing precedence