# Before

    <p>Formatted 42 queries</p>
    <p>Average query length of 280.33333 characters</p>
    <p>
        Took 0.01447 seconds total,
        0.00034 seconds per query,
        0.00123 seconds per 1000 characters
    </p>
    <p>Used 0 bytes of memory</p>

# After

    <p>Formatted 42 queries</p>
    <p>Average query length of 280.33333 characters</p>
    <p>
        Took 0.01512 seconds total,
        0.00036 seconds per query,
        0.00128 seconds per 1000 characters
    </p>
    <p>Used 0 bytes of memory</p>

There's a small performance hit, but I think we can ignore it.