@greg0ire 2020-04-19
Closing in favor of #48